+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 
 - `Bucket.remove_entry` method to remove entry from bucket, [PR-65](https://github.com/reductstore/reduct-js/pull/65)
+- `limit` option to `Bucket.query`, [PR-66](https://github.com/reductstore/reduct-js/pull/66)
 
 ## [1.5.0] - 2023-06-30